+++ Assets/Vision/EMVisionShader.shader
@@ -29,18 +29,18 @@
 
 			struct appdata
 			{
-				float4 vertex : POSITION;
-				float2 uv : TEXCOORD0;
-				float4 color : COLOR;
-				float3 normal : NORMAL;
+				half4 vertex : POSITION;
+				half2 uv : TEXCOORD0;
+				half4 color : COLOR;
+				half3 normal : NORMAL;
 			};
 
 			struct v2f
 			{
-				float2 uv : TEXCOORD0;
-				float4 vertex : SV_POSITION;
-				float3 normal : TEXCOORD1;
-				float3 viewDir : TEXCOORD2;
+				half2 uv : TEXCOORD0;
+				half4 vertex : SV_POSITION;
+				half3 normal : TEXCOORD1;
+				half3 viewDir : TEXCOORD2;
 			};
 
 			v2f vert(appdata v)
@@ -54,14 +54,14 @@
 			}
 
 			sampler2D _MainTex;
-			float4 _MainTex_TexelSize;
-			float _Coeff;
-			fixed4 _Low;
-			fixed4 _High;
+			half4 _MainTex_TexelSize;
+			half _Coeff;
+			half4 _Low;
+			half4 _High;
 
-			fixed4 frag (v2f i) : SV_Target
+			half4 frag (v2f i) : SV_Target
 			{
-				fixed4 colOrg = tex2D(_MainTex, i.uv);
+				half4 colOrg = tex2D(_MainTex, i.uv);
 
 				// filter background
 				if (colOrg.r < 0.02)
@@ -69,20 +69,20 @@
 
 
 				half d = min(1, max(0, dot(i.viewDir, i.normal)));
-				fixed4 col = d * colOrg;
+				half4 col = d * colOrg;
 
-				fixed2 off = fixed2(0.05f, 0.05f); 
+				half2 off = half2(0.05f, 0.05f); 
 				col *= _Coeff; 
-				float c2 = _Coeff * 0.5f;
+				half c2 = _Coeff * 0.5f;
 				col += tex2D(_MainTex, i.uv + (off.x / _MainTex_TexelSize.z)) * c2;
 				col += tex2D(_MainTex, i.uv - (off.y / _MainTex_TexelSize.w)) * c2;
 
-				float lum = min(3.0f, (col.r + col.g + col.b) / 3.0f);
+				half lum = min(3.0f, (col.r + col.g + col.b) / 3.0f);
 
-				float ix = (lum < 0.5f) ? 0.0 : 1.0;
+				half ix = (lum < 0.5f) ? 0.0 : 1.0;
 
-				float xSpace = _MainTex_TexelSize.z * 0.5f;
-				float ySpace = _MainTex_TexelSize.w * 0.5f;
+				half xSpace = _MainTex_TexelSize.z * 0.5f;
+				half ySpace = _MainTex_TexelSize.w * 0.5f;
 				if (frac(i.uv.x * xSpace) <= 0.5f && frac(i.uv.y * ySpace) <= 0.5f)
 					return lerp(_High, _Low, (lum - ix * 0.5f) / 0.5); //even line 
 				else
